Move document helper functions to `documentHelper` directory.
